Moving helpers into Helpers
